fix: remove ":" in filename when downloading (#172)
* fix: remove ":" in filename when downloading

In some shows (like Columbo), episodes have a colon ":" in the filename. This breaks the ffmpeg command when downloading. Which is why I changed the script to remove the character if present.
I don't know if this is the best way to do it, but it seems to work nonetheless.

* fix: refactor ":" removal

I refactored my previous fix by using tr instead of sed and printf instead of echo.

* fix: refactor printf

Refactored the printf.